# lib/api-zod/src/generated/types/createCommentBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Helpdesk API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CreateCommentBody {
  authorName: string;
  authorEmail: string;
  content: string;
  isInternal: boolean;
}