# lib/api-zod/src/generated/types/createKbArticleBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Helpdesk API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CreateKbArticleBody {
  title: string;
  content: string;
  category: string;
  tags: string[];
  published: boolean;
  /** @nullable */
  relatedTicketId?: number | null;
}