# lib/api-zod/src/generated/types/ticket.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Helpdesk API specification
 * OpenAPI spec version: 0.1.0
 */
import type { TicketPriority } from "./ticketPriority";
import type { TicketStatus } from "./ticketStatus";

export interface Ticket {
  id: number;
  title: string;
  description: string;
  status: TicketStatus;
  priority: TicketPriority;
  category: string;
  submitterName: string;
  submitterEmail: string;
  /** @nullable */
  assigneeId: number | null;
  /** @nullable */
  assigneeName: string | null;
  /** @nullable */
  resolvedAt: Date | null;
  createdAt: Date;
  updatedAt: Date;
  attachmentCount: number;
  commentCount: number;
}